# Conflicts:
#	springboot_project/ojoa/target/classes/META-INF/maven/com.example/demoJ01/pom.properties
#	springboot_project/ojoa/target/classes/com/ojo/ojoa/controller/OrdersController.class
#	springboot_project/ojoa/target/classes/com/ojo/ojoa/controller/ProductController.class
#	springboot_project/ojoa/target/classes/com/ojo/ojoa/repository/OrdersRepository.class
#	springboot_project/ojoa/target/classes/com/ojo/ojoa/service/OrdersService.class
#	springboot_project/ojoa/target/classes/com/ojo/ojoa/service/OrdersServiceImpl.class
#	springboot_project/ojoa/target/classes/com/ojo/ojoa/service/ProductsService.class

+++ springboot_project/ojoa/src/main/java/com/ojo/ojoa/repository/OrdersRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.ojo.ojoa.entity.Orders;
 
 @Repository
-public interface OrdersRepository extends JpaRepository<Orders, Long> {
+public interface OrdersRepository extends JpaRepository<Orders, Integer> {
